+++ instrumentation-api-appender/src/main/java/io/opentelemetry/instrumentation/api/appender/GlobalLogEmitterProvider.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.appender.api;
+package io.opentelemetry.instrumentation.api.appender;
 
 import io.opentelemetry.api.internal.GuardedBy;
 import javax.annotation.Nullable;
