+++ instrumentation/log4j/log4j-2.16/library/src/main/java/io/opentelemetry/instrumentation/log4j/v2_16/LogEventMapper.java
@@ -9,8 +9,8 @@ import io.opentelemetry.api.common.AttributeKey;
 import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.api.common.AttributesBuilder;
 import io.opentelemetry.context.Context;
-import io.opentelemetry.instrumentation.appender.api.LogBuilder;
-import io.opentelemetry.instrumentation.appender.api.Severity;
+import io.opentelemetry.instrumentation.api.appender.LogBuilder;
+import io.opentelemetry.instrumentation.api.appender.Severity;
 import java.util.concurrent.TimeUnit;
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.core.LogEvent;
