+++ instrumentation/log4j/log4j-2.16/library/src/test/java/io/opentelemetry/instrumentation/log4j/v2_16/OpenTelemetryAppenderConfigTest.java
@@ -12,8 +12,8 @@ import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.api.trace.Span;
 import io.opentelemetry.api.trace.SpanContext;
 import io.opentelemetry.context.Scope;
-import io.opentelemetry.instrumentation.appender.api.DelegatingLogEmitterProvider;
-import io.opentelemetry.instrumentation.appender.api.GlobalLogEmitterProvider;
+import io.opentelemetry.instrumentation.api.appender.DelegatingLogEmitterProvider;
+import io.opentelemetry.instrumentation.api.appender.GlobalLogEmitterProvider;
 import io.opentelemetry.sdk.common.InstrumentationLibraryInfo;
 import io.opentelemetry.sdk.logs.SdkLogEmitterProvider;
 import io.opentelemetry.sdk.logs.data.LogData;
